--- src/test/scala/com/thenewmotion/ochp/converters/DurationSpec.scala
@@ -1,13 +1,0 @@
-package com.thenewmotion.ochp
-package converters
-
-import converters.DurationConverter._
-
-
-class DurationSpec extends Spec {
-  "converting from Ochp and back yields the original value" >> {
-    val duration = "123:45:59"
-
-    fromOchp(duration).map(toOchp) must beSuccessfulTry(duration)
-  }
-}
